+++ lib/roo_on_rails/checks/environment.rb
@@ -19,6 +19,15 @@ module RooOnRails
       requires Documentation::Playbook
       requires Papertrail::All
 
+      def initialize(options = {})
+        super
+        @dry_run = options.fetch(:dry_run, false)
+      end
+
+      def run
+        super unless @dry_run
+      end
+
       def call
         # nothing to do
       end