+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# HEAD
+# HEAD (2017-07-24)
 
 Features:
 
 - Adds pre-baked Google OAuth support (#44, #49)
+- reports Sidekiq metrics only for queues defined in process (#50)
 - adds 'default' to list of default Sidekiq queues
 - accepts custom Sidekiq queue names
 - accepts custom Sidekiq queue permitted latency values